Update README.md
Throughout this project. I demonstrate the steps take to create an algorithm which identifies IP addresses that should no longer be able to access a designated site. This algorithm takes the IP addresses from this remove list and automatically removes them from the list of allowed addresses. With each step, I used different tools and techniques to further the progress, such as separating each item on the original file into a list and creating a loop process to search through the entire list of allowed IPs for those that are on the “remove_list” and automatically removing them. Afterwards, compiling the new data to overwrite the original files.